tuteria/requirements.txt
amqp==2.6.1
aniso8601==7.0.0
appdirs==1.4.4
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
billiard==3.5.0.5
black==21.6b0
cachetools==4.2.2
celery==4.1.0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
cloudinary==1.8.0
contentful-management==2.6.0
dataclasses==0.8
decorator==5.0.9
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.4.2
# Django==1.11.7
django-activelink==1.0
django-advanced-filters==1.0.6
django-allauth==0.37.0
# django-andablog==2.4.0
django-annoying==0.10.6
django-autocomplete-light==3.2.10
django-autoslug==1.9.3
django-bootstrap-pagination==1.6.3
django-bootstrap3==9.1.0
django-braces==1.4.0
django-cachalot==1.5.0
django-celery-beat==1.1.0
django-ckeditor==5.3.1
django-compat==1.0.15
django-configurations==2.0
django-countries==5.0
django-crispy-forms==1.7.0
django-debug-toolbar==1.9.1
django-embed-video==1.1.2
django-environ==0.4.4
django-extensions==1.9.7
# django-filter==2.4.0
django-geoposition==0.3.0
django-hijack==2.1.4
django-hijack-admin==2.1.4
django-import-export==0.6.0
django-ipware==1.1.6
django-jchart==0.4.2
django-js-asset==1.2.2
django-js-reverse==0.7.3
django-languages==0.1
django-markitup==3.7.0
django-markupfield==1.4.3
django-model-utils==3.0.0
django-pagedown==0.1.3
django-paypal==0.4.1
django-phonenumber-field==1.3.0
django-ratelimit==1.1.0
django-redis==4.8.0
django-scheduler==0.8.6.1
django-secure==1.0.1
django-static-sitemaps==4.4.0
django-taggit==0.22.1
django-test-plus==1.0.20
django-webpack-loader==0.5.0
djangorestframework==3.7.3
djangorestframework-filters==0.10.2
docopt==0.4.0
elastic-apm==6.2.3
ephem==4.0.0.2
et-xmlfile==1.1.0
execnet==1.9.0
facebook-sdk==2.0.0
factory-boy==2.9.2
Faker==8.9.0
fancycompleter==0.9.1
flower==0.9.3
future==0.16.0
geopy==1.11.0
google-auth==1.32.0
graphene==2.1.8
graphene-django==2.0.0
-e git+https://github.com/gbozee/graphene-utils.git@593e31dcb1da5311d00aee6ca48264f0730034bc#egg=graphene_utils
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==0.4.17
gunicorn==19.7.1
hiredis==0.2.0
icalendar==4.0.7
idna==2.10
importlib-metadata==4.6.0
iniconfig==1.1.1
ipython==7.16.1
ipython-genutils==0.2.0
iso8601==0.1.14
jdcal==1.4.1
jedi==0.17.2
kombu==4.1.0
mailchimp==2.0.9
Markdown==2.6.9
meinheld==1.0.2
mock==4.0.3
mypy-extensions==0.4.3
newrelic==2.98.0.81
oauthlib==3.1.1
olefile==0.46
openpyxl==2.4.9
packaging==21.0
Paperboy==1.0.1
parso==0.7.1
pathspec==0.8.1
pdbpp==0.10.2
pexpect==4.8.0
phonenumberslite==8.12.26
pickleshare==0.7.5
Pillow==4.3.0
pluggy==0.13.1
promise==2.3
prompt-toolkit==3.0.19
psycopg2==2.7.3.2
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygeoip==0.3.2
Pygments==2.9.0
pyparsing==2.4.7
pyrepl==0.9.0
pytest==6.2.4
pytest-django==4.4.0
pytest-mock==3.6.1
pytest-pythonpath==0.7
pytest-xdist==1.12
python-dateutil==2.8.1
python-google-places==1.4.1
python3-openid==3.2.0
pytz==2021.1
raven==6.10.0
redis==3.5.3
regex==2021.4.4
repoze.lru==0.7
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
Rx==1.6.1
-e git+https://github.com/gbozee/django-simple-social-login.git@65c3a6578b334d3b83807d480148ea44739f7c08#egg=simple_social_login
simplejson==3.6.5
singledispatch==3.6.2
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
text-unidecode==1.3
toml==0.10.2
tornado==5.1.1
traitlets==4.3.3
-e git+https://gitlab.com/gbozee/api-gateway-client.git@0d6b1303cc08a06161b2ff354b09601569efdba7#egg=tuteria_api_client
-e git+https://gitlab.com/gbozee/hubspot-crm.git@4c2754c2ca089f4ccf73e2d49ff4868d99cb378b#egg=tuteria_hubspot
typed-ast==1.4.3
typing==3.7.4.3
typing-extensions==3.10.0.0
unicode-slugify==0.1.3
Unidecode==1.2.0
urllib3==1.26.6
uWSGI==2.0.15
vine==1.3.0
wcwidth==0.2.5
whitenoise==3.3.1
wmctrl==0.4
zipp==3.4.1
